map to report
